+++ eve/workers/openstack-single-node-with-cypress/terraform/common.tf
@@ -1,0 +1,8 @@
+resource "random_string" "current" {
+  length  = 5
+  special = false
+}
+
+locals {
+  prefix = "metalk8s-${random_string.current.result}"
+}