+++ examples/metalk8s-solution-example/operator/pkg/controller/add_versionserver.go
@@ -1,0 +1,10 @@
+package controller
+
+import (
+	"example-solution-operator/pkg/controller/versionserver"
+)
+
+func init() {
+	// AddToManagerFuncs is a list of functions to create controllers and add them to a manager.
+	AddToManagerFuncs = append(AddToManagerFuncs, versionserver.Add)
+}