+++ examples/metalk8s-solution-example/operator/pkg/controller/controller.go
@@ -1,0 +1,20 @@
+package controller
+
+import (
+	"example-solution-operator/pkg/config"
+
+	"sigs.k8s.io/controller-runtime/pkg/manager"
+)
+
+// AddToManagerFuncs is a list of functions to add all Controllers to the Manager
+var AddToManagerFuncs []func(manager.Manager, *config.OperatorConfig) error
+
+// AddToManager adds all Controllers to the Manager
+func AddToManager(m manager.Manager, config *config.OperatorConfig) error {
+	for _, f := range AddToManagerFuncs {
+		if err := f(m, config); err != nil {
+			return err
+		}
+	}
+	return nil
+}