+++ images/metalk8s-utils/configure-repos.sh
@@ -1,0 +1,27 @@
+#!/bin/bash
+
+set -xue -o pipefail
+
+CENTOS_VERSION=$1
+SALT_VERSION=$2
+
+cat > /etc/yum.repos.d/kubernetes.repo << EOF
+[kubernetes]
+name=Kubernetes
+baseurl=https://packages.cloud.google.com/yum/repos/kubernetes-el$CENTOS_VERSION-x86_64
+enabled=1
+gpgcheck=1
+repo_gpgcheck=1
+gpgkey=https://packages.cloud.google.com/yum/doc/yum-key.gpg
+       https://packages.cloud.google.com/yum/doc/rpm-package-key.gpg
+EOF
+
+cat > /etc/yum.repos.d/saltstack.repo << EOF
+[saltstack]
+name=SaltStack repo for RHEL/CentOS \$releasever
+baseurl=https://repo.saltstack.com/yum/redhat/\$releasever/\$basearch/archive/$SALT_VERSION
+enabled=1
+gpgcheck=1
+gpgkey=https://repo.saltstack.com/yum/redhat/\$releasever/\$basearch/archive/$SALT_VERSION/SALTSTACK-GPG-KEY.pub
+       https://repo.saltstack.com/yum/redhat/\$releasever/\$basearch/archive/$SALT_VERSION/base/RPM-GPG-KEY-CentOS-$CENTOS_VERSION
+EOF